Documentation/Temporary-Codebases-for-POCs/Vikrant-Python/kafka-miniprojects/requirements.txt
absl-py==0.13.0
alembic==1.7.5
appdirs==1.4.4
astunparse==1.6.3
attrs==21.2.0
backcall==0.2.0
beautifulsoup4==4.10.0
bleach==4.1.0
cachetools==4.2.2
certifi==2021.5.30
charset-normalizer==2.0.4
clang==5.0
click==8.0.1
colorama==0.4.4
cycler==0.10.0
debugpy==1.4.1
decorator==5.0.9
defusedxml==0.7.1
entrypoints==0.3
et-xmlfile==1.1.0
Flask==1.1.1
Flask-Assets==2.0
Flask-Cors==3.0.10
Flask-Login==0.4.1
Flask-Migrate==2.5.2
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.2
flatbuffers==1.12
Fraction==2.1.1
gast==0.4.0
google-auth==1.35.0
google-auth-oauthlib==0.4.5
google-pasta==0.2.0
grpcio==1.43.0
grpcio-tools==1.43.0
gunicorn==20.1.0
h5py==3.1.0
idna==3.2
importlib-metadata==4.8.1
ipykernel==6.2.0
ipython==7.26.0
ipython-genutils==0.2.0
itsdangerous==2.0.1
jdcal==1.4.1
jedi==0.18.0
Jinja2==3.0.1
joblib==1.0.1
jsonschema==3.2.0
jupyter-client==7.0.1
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
kafka-python==2.0.2
keras==2.6.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
lxml==4.6.5
Mako==1.1.6
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.4.3
matplotlib-inline==0.1.2
mistune==0.8.4
nbclient==0.5.4
nbconvert==6.1.0
nbformat==5.1.3
nest-asyncio==1.5.1
nltk==3.6.2
notebook-as-pdf==0.5.0
numpy==1.19.5
oauthlib==3.1.1
opencv-python==4.5.5.62
openpyxl==3.0.3
opt-einsum==3.3.0
packaging==21.0
pandas==1.3.2
pandocfilters==1.5.0
parso==0.8.2
pickleshare==0.7.5
Pillow==8.3.1
prompt-toolkit==3.0.20
protobuf==3.17.3
py4j==0.10.9
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyee==8.2.2
Pygments==2.10.0
pyparsing==2.4.7
PyPDF2==1.26.0
pyppeteer==0.2.6
pyrsistent==0.18.0
pyspark==3.1.2
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2019.3
pywin32==301
pyzmq==22.2.1
regex==2021.8.21
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
scipy==1.7.2
siphon==0.9
six==1.14.0
soupsieve==2.3.1
SQLAlchemy==1.3.13
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.6.0
tensorflow-estimator==2.6.0
termcolor==1.1.0
testpath==0.5.0
tornado==6.1
tqdm==4.62.1
traitlets==5.0.5
typing-extensions==3.7.4.3
urllib3==1.26.6
wcwidth==0.2.5
webassets==2.0
webencodings==0.5.1
websockets==9.1
Werkzeug==0.16.0
wrapt==1.12.1
WTForms==2.2.1
xlrd==1.2.0
-e git+ssh://git@github.com/vikrantdeshpande09876/XML_to_CSV_Rep.git@94bfe3697c528f098e022e64b4bd4f41c90445f3#egg=XML2CSV
zipp==3.5.0
